+++ apps/web/src/components/tx-flow/actions/Execute/styles.module.css
@@ -1,7 +1,3 @@
-.params {
-  margin-bottom: var(--space-2);
-}
-
 .noBottomBorderRadius :global(.MuiPaper-root) {
   border-bottom-left-radius: 0 !important;
   border-bottom-right-radius: 0 !important;
