+++ apps/web/src/components/tx-flow/actions/ExecuteThroughRole/ExecuteThroughRoleForm/__test__/ExecuteThroughRoleForm.test.tsx
@@ -12,7 +12,7 @@ import * as txSender from '@/services/tx/tx-sender/dispatch'
 import { extendedSafeInfoBuilder } from '@/tests/builders/safe'
 import { type OnboardAPI } from '@web3-onboard/core'
 import { AbiCoder, ZeroAddress, encodeBytes32String } from 'ethers'
-import ExecuteThroughRoleForm from '..'
+import ExecuteThroughRoleForm from '../index'
 import * as hooksModule from '../hooks'
 import { chainBuilder } from '@/tests/builders/chains'
 import { useHasFeature } from '@/hooks/useChains'
