+++ apps/web/src/components/tx-flow/actions/ExecuteThroughRole/ExecuteThroughRoleForm/styles.module.css
@@ -1,0 +1,11 @@
+.roleChip {
+  background-color: var(--color-background-main);
+  font-size: 12px;
+  border-radius: 4px;
+  padding: 2px 8px;
+}
+
+.zodiac {
+  color: var(--color-text-primary);
+}
+
