+++ apps/web/src/components/tx-flow/actions/ExecuteThroughRole/index.tsx
@@ -1,9 +1,9 @@
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { useCallback, useContext } from 'react'
-import { TxFlowContext } from '../TxFlowProvider'
-import ExecuteThroughRoleForm from '@/components/tx/SignOrExecuteForm/ExecuteThroughRoleForm'
+import { TxFlowContext } from '../../TxFlowProvider'
+import ExecuteThroughRoleForm from './ExecuteThroughRoleForm'
 import useIsCounterfactualSafe from '@/features/counterfactual/hooks/useIsCounterfactualSafe'
-import { type SlotComponentProps, SlotName, withSlot } from '../slots'
+import { type SlotComponentProps, SlotName, withSlot } from '../../slots'
 
 const ExecuteThroughRole = ({ onSubmit }: SlotComponentProps<SlotName.Submit>) => {
   const { safeTx } = useContext(SafeTxContext)
