+++ apps/web/src/components/tx-flow/common/styles.module.css
@@ -16,3 +16,7 @@
 .form > :global(.MuiFormControl-root) {
   margin-bottom: 28px;
 }
+
+.params {
+  margin-bottom: var(--space-2);
+}
