+++ apps/web/src/components/tx/SignOrExecuteForm/__tests__/SignOrExecuteFormV2.test.tsx
@@ -1,5 +1,5 @@
 import * as hooks from '@/components/tx/SignOrExecuteForm/hooks'
-import * as execThroughRoleHooks from '@/components/tx/SignOrExecuteForm/ExecuteThroughRoleForm/hooks'
+import * as execThroughRoleHooks from '@/components/tx-flow/actions/ExecuteThroughRole/ExecuteThroughRoleForm/hooks'
 import { safeTxBuilder } from '@/tests/builders/safeTx'
 import { render } from '@/tests/test-utils'
 import type { TransactionDetails } from '@safe-global/safe-gateway-typescript-sdk'
